Add type variable to type [path]
The type `path` now has three type variables:
- the `kind` of the path (absolute or relative)
- its kind once resolved
- the `object` identified by the path

When concatenating two paths, we used a `cons` type to encode the
relation between the kind of concatenated paths and the result of
concatenation. Now we have to use a second `cons` value to encode the
same constraints on resolved kinds.

I introduced a new GADT called `impl` to implement the fact that if a
path is absolute, its resolved path is absolute too.

The main benefit of adding one type variable is that the return type
of `resolve` can be fully derived from the type of its arguments.

Under this representation however, I think writing the type of
`parent` is not possible. Taking the parent of a path may change the
kind of the resolved path or not, depending on the first item of the
path (basically, if it's a link pointing to an absolute path).