New behaviour of {rel,abs} * abs concatenation
Adopted a new rule for concatenation when the second argument is
absolute. The idea is to mimic what happens intuitively at the
"syntactic level". E.g.

/foo + /bar = /foo/bar
.. + /foo = ../foo

One consequence is that the type of `concat` simplifies, as the kind
of the result is the kind of the first argument, while the object of
the result still is the object of the right argument. In consequence
there is no more need for the `cons` (may it rest in peace).